refactoring for namespace group perms
